Second iteration of search
