+++ src/lib/components/loadingPage.svelte
@@ -1,0 +1,8 @@
+<script>
+    import jumboCat from '$lib/assets/jumbo-cat.gif';
+</script>
+
+<div class="w-full h-screen flex flex-col gap-2 items-center justify-center">
+    <p class="text-3xl">Loading...</p>
+    <img src={jumboCat} alt="Happy Cat" width="40%" height="40%" />
+</div>
