+++ src/routes/search/+layout.svelte
@@ -17,7 +17,7 @@
 </script>
 
 <div class="w-full flex sticky top-0 backdrop-blur-md justify-center items-center px-4">
-	<a href="/">
+	<a href="/" aria-label="Back to Homepage">
 		<svg
 			xmlns="http://www.w3.org/2000/svg"
 			fill="none"
